requirements.txt
anthropic
google-auth
google-auth-oauthlib
google-api-python-client
html2text
